A disguisting amalgmation of code, just so Andy can see it
Just a small-town girl, livin' in a lonely world...
