+++ src/WooCommerce/class-ajax.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @see WC_AJAX::do_wc_ajax()
+ */
+namespace TripleA_Payment_Gateway_For_WooCommerce\WooCommerce;
+
+/**
+ * Class AJAX
+ *
+ * @package TripleA_Payment_Gateway_For_WooCommerce\WooCommerce
+ */
+class AJAX {
+
+}
