+++ src/WooCommerce/index.php
