+++ src/includes/index.php
@@ -1,1 +1,8 @@
-<?php // Silence is golden
+<?php
+/**
+ * Silence is golden.
+ *
+ * @package    TripleA_Payment_Gateway_For_WooCommerce
+ */
+
+die();
