+++ tests/integration/public/class-public-Test.php
@@ -1,0 +1,1 @@
+<?php
